+++ integration_tests/tests/ds.rs
@@ -28,28 +28,28 @@ fn should_describe_directories() {
 fn should_conditional_forwarders() {
     let client = DirectoryServiceClient::new(Region::UsWest2);
     let mut request = DescribeConditionalForwardersRequest::default();
-	request.directory_id = "d-11111aaaaa".to_string();
+    request.directory_id = "d-11111aaaaa".to_string();
 
     match client.describe_conditional_forwarders(request).sync() {
-		  Err(RusotoError::Service(DescribeConditionalForwardersError::EntityDoesNotExist(msg))) => {
+        Err(RusotoError::Service(DescribeConditionalForwardersError::EntityDoesNotExist(msg))) => {
             assert!(msg.contains("does not exist."))
         }
         err @ _ => panic!("Expected EntityDoesNotExist error, got {:#?}", err)
-	}
+    }
 }
 
 #[test]
 fn should_describe_domain_controllers() {
     let client = DirectoryServiceClient::new(Region::UsWest2);
     let mut request = DescribeDomainControllersRequest::default();
-	request.directory_id = "d-11111aaaaa".to_string();
+    request.directory_id = "d-11111aaaaa".to_string();
 
     match client.describe_domain_controllers(request).sync() {
-		Err(RusotoError::Service(DescribeDomainControllersError::EntityDoesNotExist(msg))) => {
+        Err(RusotoError::Service(DescribeDomainControllersError::EntityDoesNotExist(msg))) => {
             assert!(msg.contains("does not exist."))
         }
         err @ _ => panic!("Expected EntityDoesNotExist error, got {:#?}", err)
-	}
+    }
 }
 
 #[test]